Updated the license for the project
As the project continues to grow in size, and more dedication and time is put into the project. Having the necessary copyright license in place early is a godsend for any issues that may appear later down the lines as per experience from past projects. If you feel this decision or license is wrong, please reach out to me.